requirements.txt
# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Data validation
pydantic>=2.5.0
pydantic[email]>=2.5.0

# MongoDB
pymongo>=4.6.0
motor>=3.3.0

# Environment variables
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0

